Add transition exception to change control to new interpreter instance
In the ModelTransitionValue when calling transition, raise a special
exception InterpreterTransitionException including the name of the model
to transition to. In InterpreterCmd, if the interpreter instance throws
InterpreterTransitionException then call self and start new interpreter
instance with the file name supplied from the exception.

Somewhat an anti-pattern to use exceptions for control flow, but until
we come up with a better way ...

Signed-off-by: Henrik Ejersbo <hejersbo@ece.au.dk>